moving testing_test to assert_test
